Merge remote-tracking branch 'origin/master' into develop
# Conflicts:
#	slackbot/dispatcher.py
The settings are passed through the plugin manager in the init. Removed
some useless imports. Message matcher inside the object so that the
alias’ are loaded from the settings passed down through the bot start,
not through an import statement. (Can move the settings file to other
locations for different implementations of updating the bot)